chore: refactor selected suite hooks and utils
