+++ packages/suite/src/components/suite/modals/Passphrase/components/PassphraseTypeCard/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useRef, useEffect, useCallback } from 'react';
 import { AnimatePresence, motion } from 'framer-motion';
 import { ANIMATION } from '@suite-config';
-import { setCaretPosition, useKeyPress } from '@suite-utils/dom';
+import { setCaretPosition } from '@suite-utils/dom';
 import styled, { css } from 'styled-components';
 import { Button, useTheme, variables, Input, Tooltip, Checkbox, Icon } from '@trezor/components';
 import { Translation } from '@suite-components/Translation';
@@ -9,7 +9,7 @@ import { MAX_LENGTH } from '@suite-constants/inputs';
 import { countBytesInString } from '@trezor/utils';
 import { OpenGuideFromTooltip } from '@guide-components';
 import PasswordStrengthIndicator from '@suite-components/PasswordStrengthIndicator';
-import { useTranslation } from '@suite-hooks';
+import { useTranslation, useKeyPress } from '@suite-hooks';
 import { isAndroid } from '@suite-utils/env';
 
 const Wrapper = styled.div<Pick<Props, 'type' | 'singleColModal'>>`
