+++ packages/suite/src/hooks/suite/useKeyPress.ts
@@ -1,0 +1,37 @@
+import { useState, useEffect } from 'react';
+
+// todo: duplicity
+export const useKeyPress = (targetKey: string) => {
+    // State for keeping track of whether key is pressed
+    const [keyPressed, setKeyPressed] = useState(false);
+
+    // If pressed key is our target key then set to true
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    const downHandler = (event: KeyboardEvent) => {
+        if (event.key === targetKey) {
+            setKeyPressed(true);
+        }
+    };
+
+    // If released key is our target key then set to false
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+    const upHandler = (event: KeyboardEvent) => {
+        if (event.key === targetKey) {
+            setKeyPressed(false);
+        }
+    };
+
+    // Add event listeners
+    useEffect(() => {
+        window.addEventListener('keydown', downHandler);
+        window.addEventListener('keyup', upHandler);
+        // Remove event listeners on cleanup
+        return () => {
+            window.removeEventListener('keydown', downHandler);
+            window.removeEventListener('keyup', upHandler);
+        };
+        // eslint-disable-next-line react-hooks/exhaustive-deps
+    }, []); // Empty array ensures that effect is only run on mount and unmount
+
+    return keyPressed;
+};
